Database clustering works

+++ mattermost_setup.sh
@@ -7,9 +7,15 @@ mkdir -p /media/mmst-data
 cat /vagrant/client_fstab >> /etc/fstab
 mount -a
 
-cp /vagrant/mattermost-5.9.0-linux.amd64.tar.gz ./
-tar -xzf mattermost*.gz
+# Download Mattermost
+if [ ! -f /vagrant/mattermost-5.9.0-linux.amd64.tar.gz ]; then
+    wget --quiet https://releases.mattermost.com/5.9.0/mattermost-5.9.0-linux-amd64.tar.gz
+    cp mattermost-5.9.0-linux-amd64.tar.gz /vagrant/mattermost-5.9.0-linux.amd64.tar.gz
+else
+	cp /vagrant/mattermost-5.9.0-linux.amd64.tar.gz ./	
+fi
 
+tar -xzf mattermost*.gz
 rm mattermost*.gz
 mv mattermost /opt
 
@@ -17,15 +23,11 @@ mkdir /opt/mattermost/data
 
 ln -s /vagrant/license.txt /opt/mattermost/license.txt
 mv /opt/mattermost/config/config.json /opt/mattermost/config/config.orig.json
-jq '.ServiceSettings.LicenseFileLocation = "/opt/mattermost/license.txt"' /opt/mattermost/config/config.orig.json > /vagrant/config.tmp.json && mv /vagrant/config.tmp.json /vagrant/config.json
-jq '.SqlSettings.DataSource = "mmuser:really_secure_password@tcp(192.168.33.101:3306)/mattermost?charset=utf8mb4,utf8\u0026readTimeout=30s\u0026writeTimeout=30s"' /vagrant/config.json > /vagrant/config.tmp.json && mv /vagrant/config.tmp.json /vagrant/config.json
-jq '.ClusterSettings.Enable = true' /vagrant/config.json > /vagrant/config.tmp.json && mv /vagrant/config.tmp.json /vagrant/config.json
-jq '.ClusterSettings.ClusterName = "Buster"' /vagrant/config.json > /vagrant/config.tmp.json && mv /vagrant/config.tmp.json /vagrant/config.json
-jq '.ClusterSettings.OverrideHostname = "#IP_ADDR"' /vagrant/config.json > /vagrant/config.tmp.json && mv /vagrant/config.tmp.json /vagrant/config.json
-jq '.ClusterSettings.ReadOnlyConfig = false' /vagrant/config.json > /vagrant/config.tmp.json && mv /vagrant/config.tmp.json /vagrant/config.json
-jq '.FileSettings.Directory = "/media/mmst-data/"' /vagrant/config.json > /vagrant/config.tmp.json && mv /vagrant/config.tmp.json /vagrant/config.json
 
-cp /vagrant/config.json /opt/mattermost/config/config.json
+jq -s '.[0] * .[1]' /opt/mattermost/config/config.orig.json /vagrant/config.json > /vagrant/config.vagrant.json
+chmod 777 /vagrant/config.vagrant.json
+
+ln -s /vagrant/config.vagrant.json /opt/mattermost/config/config.json
 
 useradd --system --user-group mattermost
 chown -R mattermost:mattermost /opt/mattermost
